+++ Cargo.toml
@@ -57,6 +57,7 @@ egui = "0.31"
 egui-wgpu = "0.31"
 eframe = { version = "0.31", default-features = false }
 env_logger = "0.11"
+font-kit = "0.14"
 font-query = { git = "https://github.com/Maximkaaa/font-query" }
 futures = "0.3"
 futures-intrusive = "0.5"
