+++ galileo/Cargo.toml
@@ -28,6 +28,7 @@ base64 = { workspace = true }
 bytemuck = { workspace = true, features = ["derive"] }
 bytes = { workspace = true }
 cfg-if = { workspace = true }
+font-kit = { workspace = true }
 font-query = { workspace = true }
 futures-intrusive = { workspace = true }
 galileo-mvt = { workspace = true }
